utime_t cleanup

+++ src/config.cc
@@ -49,8 +49,8 @@ int _get_bits_of(int v) {
   return n;
 }
 unsigned _page_size = sysconf(_SC_PAGESIZE);
-unsigned _page_mask = ~(_page_size - 1);
-unsigned long _page_shift = _get_bits_of(_page_size);
+unsigned long _page_mask = ~(_page_size - 1);
+unsigned _page_shift = _get_bits_of(_page_size);
 
 // file layouts
 struct ceph_file_layout g_OSD_FileLayout = {
